Parametrizing data for the database in the WebLab instance creation and some related issues.
The host is selected between the first of the hosts in DB group (for sake of clarity, let's suppose that it will only be one).
The database name, user and password are specified in the global variable file because they are needed both in the server and the client hosts.
The password should be changed in the final installation and "vaulted" if it is versioned afterwards.